requirements.txt
anthropic
flask
pdfplumber
fpdf2
python-dotenv
